+++ dec/src/decimal32.rs
@@ -105,6 +105,25 @@ impl Decimal32 {
         self.inner.bytes
     }
 
+    /// Computes the coefficient of the number.
+    ///
+    /// If the number is a special value (i.e., NaN or infinity), returns zero.
+    pub fn coefficient(&self) -> i32 {
+        let mut buf = MaybeUninit::<[u8; decnumber_sys::DECSINGLE_Pmax]>::uninit();
+        let sign = unsafe {
+            decnumber_sys::decSingleGetCoefficient(&self.inner, buf.as_mut_ptr() as *mut u8)
+        };
+        let buf = unsafe { buf.assume_init() };
+        let mut coeff = 0;
+        for n in buf.iter().skip_while(|x| **x == 0) {
+            coeff = coeff * 10 + i32::from(*n);
+        }
+        if sign < 0 {
+            coeff *= -1;
+        }
+        coeff
+    }
+
     /// Computes the exponent of the number.
     pub fn exponent(&self) -> i32 {
         unsafe { decnumber_sys::decSingleGetExponent(&self.inner) }